Added tests for BBCodeManager + bug fix
Two of the tests that were added allowed me to notice the regexp was
incomplete. Now, both '=' and ']' are forbidden in the first part of
the opening tag, and both '[' and ']' in the closing tag.

Actually, thinking about it, a couple more tests with '[' or ']' inside
the tags are in order. I will see to it later.