Added a code compressor, and a couple of tests
Now a minified version of the code is generated. The compressor used to
write warnings about a JSLint rule I had deactivated (single var per
scope). Seeing as it does not necessarily makes the code harder to
understand, I activated it back and updated the code accordingly.

I added a couple of additional test cases to reflect my remark in the
previous commit.