Added more tests
Also changed the code convention in use. Only applyTo left to be
tested, and that will be it. We will probably rely on Mootools a lot
for this last set of tests.

Then I can go back to adding Maven plugins to the project (for the code
documentation, for example).