Added tests for the applyTo method
Tests have been added to cover the uses of the applyTo method;
unfortunately, they fail. It seems to be because Jasmine cannot compare
the objects well. I'll be updating them to compare plain old strings.

Also, fixed a bug in the applyTo method. I really, really need to read
proper documentation about the management of the 'this' variable.