There is no I/O parameter in Javascript
That's where my mistake was: I was expecting the altering of an input
parameter to be reflected in the calling context. I do not know if that
was the exact cause of error at first, when everything was pretty much
at the same place, but that clearly was the case after I put the part
about string replacement in its own method. All 44 tests pass! I should
make a couple of example pages — though considering the simplicity of
the ext-* projects, it is not really needed. We will see if there are
some things that do not work soon...